+++ examples/nextjs-with-typescript/pages/MuxUploader.tsx
@@ -26,27 +26,24 @@ function MuxUploaderPage() {
       <Head>
         <title>&lt;MuxUploader/&gt; Demo</title>
       </Head>
-      <MuxUploaderDrop mux-uploader="uploader" overlay overlayText="You're doing great!">
-        <h2>Enter your upload GCS url:</h2>
+      <h2>Enter your upload GCS url:</h2>
 
-        <input type="text" style={{
-          marginBottom: "20px",
-          width: "min(100%, 400px)",
-          boxSizing: "border-box"
-        }} placeholder="https://storage.googleapis.com/..." onChange={handleChange} />
+      <input type="text" style={{
+        marginBottom: "20px",
+        width: "min(100%, 400px)",
+        boxSizing: "border-box"
+      }} placeholder="https://storage.googleapis.com/..." onChange={handleChange} />
 
-        <MuxUploader
-          style={{ display: "block", height: "200px" }}
-          id="uploader"
-          endpoint={url}
-          onUploadStart={onUploadStart}
-          onChunkAttempt={onChunkAttempt}
-          onChunkSuccess={onChunkSuccess}
-          onSuccess={onSuccess}
-          onUploadError={onUploadError}
-          onProgress={onProgress}
-        />
-      </MuxUploaderDrop>
+      <MuxUploader
+        id="uploader"
+        endpoint={url}
+        onUploadStart={onUploadStart}
+        onChunkAttempt={onChunkAttempt}
+        onChunkSuccess={onChunkSuccess}
+        onSuccess={onSuccess}
+        onUploadError={onUploadError}
+        onProgress={onProgress}
+      />
 
       <Link href="/"><a>Browse Elements</a></Link>
     </>