+++ packages/mux-uploader/src/layouts/block.ts
@@ -14,19 +14,19 @@ export default function blockLayout(contextElement: MuxUploaderElement): Documen
   const progressElements = conditionalRender(
     noProgress,
     `
-      <mux-uploader-progress type="percentage"></mux-uploader-progress>
-      <mux-uploader-progress></mux-uploader-progress>
+      <mux-uploader-progress part="progress progress-percentage" type="percentage"></mux-uploader-progress>
+      <mux-uploader-progress part="progress progress-bar" type="bar"></mux-uploader-progress>
     `
   );
-  const statusElement = conditionalRender(noStatus, '<mux-uploader-status></mux-uploader-status>');
-  const retryElement = conditionalRender(noRetry, '<mux-uploader-retry></mux-uploader-retry>');
+  const statusElement = conditionalRender(noStatus, '<mux-uploader-status part="status"></mux-uploader-status>');
+  const retryElement = conditionalRender(noRetry, '<mux-uploader-retry part="retry"></mux-uploader-retry>');
 
   return document.createRange().createContextualFragment(`
     <${wrapper}>
       ${statusElement}
       ${retryElement}
 
-      <mux-uploader-file-select>
+      <mux-uploader-file-select part="file-select">
         <slot name="file-select">
           ${fileSelectFragment}
         </slot>