+++ packages/mux-video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.2](https://github.com/muxinc/elements/compare/@mux/mux-video@0.15.1...@mux/mux-video@0.15.2) (2023-05-08)
+
+**Note:** Version bump only for package @mux/mux-video
+
+
+
+
+
 ## [0.15.1](https://github.com/muxinc/elements/compare/@mux/mux-video@0.15.0...@mux/mux-video@0.15.1) (2023-04-24)